faang_gsoc/requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
amqp==5.1.1
aniso8601==7.0.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.15
django-channels-graphql-ws==0.9.1
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.5
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
elasticsearch==7.17.0
elasticsearch-dsl==7.4.0
flower==1.2.0
frozenlist==1.3.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-ws==0.4.4
hiredis==2.0.0
humanize==4.3.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
lxml==4.9.1
msgpack==1.0.4
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tornado==6.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
yarl==1.8.1
zope.interface==5.4.0
